Some aspects of the DRBL work :-)
Widened the screen a bit :-).

Commitment store can now prove a fact, even when a rule has multiple conditions or if multiple rules apply.

Fixed a bug in the prove_conclusion function :-).

Added a function to the Commitment Store class that can check whether a given sentence is provable, given the current
facts and rules. If it can be proven, the rules that are used for the proof are returned.

Rules now have properties (valid, applicable, apply, excluded).

Forget Pyke. I'm implementing my own very simple logic, which should be sufficient for this project :-). We have Fact, Rule and CS classes. The CS (commitment store) contains a dictionary of Facts and Rules. CS will also get functions for proving whether or not a rule is a reason for a certain conclusion (forward-chaining) or the other way around (backward-chaining). These functions still have to be implemented, though....

Merge branch 'development' of https://github.com/Vishengel/dmas into development

# Conflicts:
#	Controller.py
#	Model.py

Added agent method and additional fields.